moved utils functions to wikilink.py as import is failing
