+++ wikilink/__init__.py
@@ -9,10 +9,9 @@
     :license: Apache License 2.0.
 """
 from .wiki_link import WikiLink
-from .wiki_link import utils
 
 name = "wikilink"
-__all__ = ["WikiLink", "utils"]
+__all__ = ["WikiLink"]
 __author__ = "Tran Ly Vu (vutransingapore@gmail.com)"
 __copyright__ = "Copyright (c) 2016 - 2019 Tran Ly Vu. All Rights Reserved."
 __credits__ = ["Tranlyvu"]
